+++ udata/core/badges/commands.py
@@ -12,8 +12,8 @@ from udata.models import Organization
 log = logging.getLogger(__name__)
 
 
-@cli.group()
-def badges():
+@cli.group('badges')
+def grp():
     '''Badges related operations'''
 
 
@@ -29,7 +29,7 @@ def toggle_badge(id_or_slug, badge_kind):
     organization.toggle_badge(badge_kind)
 
 
-@badges.command()
+@grp.command()
 @click.argument('path_or_id')
 @click.argument('badge_kind')
 def toggle(path_or_id, badge_kind):