+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Current (in progress)
 
 - Fix organization admin pagination [#1372](https://github.com/opendatateam/udata/issues/1372)
-- Switch to `flask-cli` and drop `flask-script`. Deprecated commands have been removed. [#1364](https://github.com/opendatateam/udata/pull/1364)
+- Switch to `flask-cli` and drop `flask-script`. Deprecated commands have been removed. [#1364](https://github.com/opendatateam/udata/pull/1364) [breaking]
 
 ## 1.2.10 (2018-01-24)